Legger til felter for ekskludert innhold og locale fallback på oversiktssider. Refactor funksjoner for å bygge oversiktslister
