+++ src/main/resources/lib/guillotine/schema/schema-creation-callbacks/forms-overview-data-callback.ts
@@ -4,8 +4,8 @@ import { CreationCallback, graphQlCreateObjectType } from '../../utils/creation-
 import { logger } from '../../../utils/logging';
 import { forceArray } from '../../../utils/array-utils';
 import { getGuillotineContentQueryBaseContentId } from '../../utils/content-query-context';
-import { buildFormDetailsList } from '../../../forms-overview/form-details-list-builder';
-import { FormDetailsListItem } from '../../../forms-overview/types';
+import { buildFormDetailsList } from '../../../overview-pages/forms-overview/form-details-list-builder';
+import { FormDetailsListItem } from '../../../overview-pages/forms-overview/types';
 
 export const formsOverviewDataCallback: CreationCallback = (context, params) => {
     const formDetailsList = graphQlCreateObjectType<keyof FormDetailsListItem>(context, {
