+++ src/main/resources/lib/overview-pages/forms-overview/form-details-list-builder.ts
@@ -1,14 +1,14 @@
 import * as contentLib from '/lib/xp/content';
-import { forceArray, removeDuplicatesFilter } from '../utils/array-utils';
-import { FormsOverview } from '../../site/content-types/forms-overview/forms-overview';
-import { contentTypesWithFormDetails } from '../contenttype-lists';
+import { forceArray, removeDuplicatesFilter } from '../../utils/array-utils';
+import { FormsOverview } from '../../../site/content-types/forms-overview/forms-overview';
+import { contentTypesWithFormDetails } from '../../contenttype-lists';
 import { ContentWithFormDetails, FormDetailsListItem, FormDetailsMap } from './types';
 import { formsOverviewListItemTransformer } from './list-item-transformer';
-import { isContentLocalized } from '../localization/locale-utils';
-import { ContentDataLocaleFallback } from '../../site/content-types/content-data-locale-fallback/content-data-locale-fallback';
-import { logger } from '../utils/logging';
-import { getLocaleFromContext } from '../localization/locale-context';
-import { sortByLocaleCompareOnField } from '../utils/sort-utils';
+import { isContentLocalized } from '../../localization/locale-utils';
+import { ContentDataLocaleFallback } from '../../../site/content-types/content-data-locale-fallback/content-data-locale-fallback';
+import { logger } from '../../utils/logging';
+import { getLocaleFromContext } from '../../localization/locale-context';
+import { sortByLocaleCompareOnField } from '../../utils/sort-utils';
 
 type Audience = FormsOverview['audience'];
 type OverviewType = FormsOverview['overviewType'];
