+++ src/main/resources/lib/overview-pages/overview/get-product-pages-for-overview.ts
@@ -1,0 +1,59 @@
+import contentLib from '/lib/xp/content';
+import {
+    contentTypesInAllProductsPage,
+    contentTypesInOverviewPages,
+} from '../../contenttype-lists';
+import { forceArray } from '../../utils/array-utils';
+import { Overview } from '../../../site/content-types/overview/overview';
+
+type Args = {
+    overviewType: Overview['overviewType'];
+    audience: Overview['audience'];
+};
+
+export const getProductPagesForOverview = ({ overviewType, audience }: Args) => {
+    const isAllProductsType = overviewType === 'all_products';
+
+    return contentLib.query({
+        start: 0,
+        count: 1000,
+        contentTypes: isAllProductsType
+            ? contentTypesInAllProductsPage
+            : contentTypesInOverviewPages,
+        filters: {
+            boolean: {
+                must: [
+                    {
+                        hasValue: {
+                            field: 'data.audience._selected',
+                            values: forceArray(audience),
+                        },
+                    },
+                    ...(!isAllProductsType
+                        ? [
+                              {
+                                  exists: {
+                                      field: `data.${overviewType}`,
+                                  },
+                              },
+                          ]
+                        : []),
+                ],
+                mustNot: [
+                    {
+                        hasValue: {
+                            field: 'x.no-nav-navno.previewOnly.previewOnly',
+                            values: [true],
+                        },
+                    },
+                    {
+                        hasValue: {
+                            field: 'data.hideFromProductlist',
+                            values: [true],
+                        },
+                    },
+                ],
+            },
+        },
+    }).hits;
+};
